Move LabelAttr to a separate attrs.py file and remove filecheck test (separate PR to follow)
