--- tests/filecheck/dialects/arm_cf/test_ops.mlir
@@ -1,9 +1,0 @@
-// RUN: XDSL_ROUNDTRIP
-// RUN: XDSL_GENERIC_ROUNDTRIP
-// RUN: xdsl-opt -t arm-asm %s | filecheck %s --check-prefix=CHECK-ASM
-
-module {
-  // CHECK: arm.label "loop"
-  ^loop:
-    arm.label "loop"
-}
